Fix a bug when removing mapping tables

+++ qtools_sxzq/utility/rm_tqdb.py
@@ -1,4 +1,5 @@
 import argparse
+import re
 from transmatrix.data_api import Database
 from qtools_sxzq.qwidgets import SFG
 
@@ -39,9 +40,12 @@ def main():
         tabs = db.show_tables()
         if tabs:
             for i, tab in enumerate(tabs):
-                print(f"removing {i:>3d} {tab}")
-                db.truncate_table(tab)
-                db.delete_table(tab)
+                if re.match(pattern=r".*_mapping_\d{13}_\d{2}$", string=tab):
+                    print(f"skip {i:>3d} {tab}, mapping file will be removed automatically when its master is removed")
+                else:
+                    print(f"removing {i:>3d} {tab}")
+                    db.truncate_table(tab)
+                    db.delete_table(tab)
         else:
             print(f"{SFG(lib_name)} has no tables")
     else:
